refactor(auth): conform lib/auth/ to naming standards
- redirectValidation.ts → utils.ts (pure utility)
- requireAuth.ts merged into svcSS.ts (server-side service); AuthCheckResult
  made package-internal (unexported)
- Update all 8 import sites
