+++ web/src/app/app/shared/[chatId]/page.tsx
@@ -1,7 +1,7 @@
 import { fetchSS } from "@/lib/utilsSS";
 import { redirect } from "next/navigation";
 import type { Route } from "next";
-import { requireAuth } from "@/lib/auth/requireAuth";
+import { requireAuth } from "@/lib/auth/svcSS";
 import SharedChatDisplay from "@/app/app/shared/[chatId]/SharedChatDisplay";
 import { Agent } from "@/lib/agents/types";
