+++ web/src/app/ee/agents/stats/[id]/page.tsx
@@ -2,7 +2,7 @@ import { InstantSSRAutoRefresh } from "@/components/SSRAutoRefresh";
 import { unstable_noStore as noStore } from "next/cache";
 import { redirect } from "next/navigation";
 import type { Route } from "next";
-import { requireAuth } from "@/lib/auth/requireAuth";
+import { requireAuth } from "@/lib/auth/svcSS";
 import { AgentStats } from "./AgentStats";
 import BackButton from "@/refresh-components/buttons/BackButton";
