+++ web/src/layouts/craft/CraftManageLayout.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import type { Route } from "next";
-import { requireAuth } from "@/lib/auth/requireAuth";
+import { requireAuth } from "@/lib/auth/svcSS";
 import { UserRole } from "@/lib/types";
 
 interface CraftManageLayoutProps {
